Test confirms that the max sample size isn't greater than the possible maximum so the bug must be elsewhere (in the processing after reading and storing the data)

+++ src/audio/file/AudioFile.java
@@ -9,7 +9,7 @@ public interface AudioFile {
     int getChannels();
     int getNumberOfSamples();
     int[] getAllSamples(int channel);
-    short[] getChunk(int samples,int channel);
+    int[] getChunk(int samples,int channel);
     int[] getSamples(double seconds, int channel);
     
     void resetPos();
